+++ bhmdoc/src/lib.rs
@@ -106,6 +106,7 @@
 //!         device_response,
 //!         current_time,
 //!         "example mdoc generated nonce",
+//!         None,
 //!         get_signature_verifier,
 //!     )
 //!     .unwrap();